Update Risk Analysis
Drawing on the results of the threat identification and vulnerability assessment, the risk analysis phase adopts the FAIR methodology. Employing quantitative analysis, this stage estimates the likelihood of threats exploiting vulnerabilities and calculates the potential impacts of the threats. This analysis will enhance the comprehension of overall risk exposure in the cloud environment.